refactor(oracles/sdk): Move capabilities utils to sdk
