+++ libs/sdk/src/oracle.rs
@@ -1,5 +1,8 @@
 use serde::{Deserialize, Serialize};
 
+pub mod capabilities;
+pub use capabilities::{get_api_keys, get_capabilities_from_settings, Capabilities};
+
 #[derive(Clone, Debug)]
 pub struct DataFeedSetting {
     pub id: String,
